Fix BuildCraft and add config option to shut up fix detector

+++ minecraft/de/thexxturboxx/blockhelper/integration/BuildcraftIntegration.java
@@ -29,7 +29,7 @@ public class BuildcraftIntegration extends BlockHelperInfoProvider {
                 Number energyStored = getField(Engine.class, engine, "energy");
                 Number maxEnergy = getField(Engine.class, engine, "maxEnergy");
                 if (energyStored != null && maxEnergy != null && maxEnergy.doubleValue() != 0) {
-                    info.add(energyStored + " MJ / " + maxEnergy + " MJ");
+                    info.add(energyStored.intValue() + " MJ / " + maxEnergy.intValue() + " MJ");
                 }
             }
         } else if (iof(state.te, "buildcraft.api.IPowerReceptor")) {
@@ -38,14 +38,14 @@ public class BuildcraftIntegration extends BlockHelperInfoProvider {
                 Number energyStored = getField(PowerProvider.class, prov, "energyStored");
                 Number maxEnergy = getField(PowerProvider.class, prov, "maxEnergyStored");
                 if (energyStored != null && maxEnergy != null && maxEnergy.doubleValue() != 0) {
-                    info.add(energyStored + " MJ / " + maxEnergy + " MJ");
+                    info.add(energyStored.intValue() + " MJ / " + maxEnergy.intValue() + " MJ");
                 }
             }
         } else if (iof(state.te, "buildcraft.api.power.IPowerReceptor")) {
             buildcraft.api.power.IPowerProvider prov =
                     ((buildcraft.api.power.IPowerReceptor) state.te).getPowerProvider();
             if (prov != null && prov.getMaxEnergyStored() != 0) {
-                info.add(prov.getEnergyStored() + " MJ / " + prov.getMaxEnergyStored() + " MJ");
+                info.add((int) prov.getEnergyStored() + " MJ / " + prov.getMaxEnergyStored() + " MJ");
             }
         }
         if (iof(state.te, "buildcraft.api.ILiquidContainer")) {
